resume_analyzer/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic-settings==2.2.1
pydantic[email]==2.6.1
pymupdf==1.23.22
python-docx==1.1.0
openai==1.13.3
fastapi-mail==1.4.1
python-dotenv==1.0.1
